+++ libs/core/src/lib/createMock.types.ts
@@ -3,15 +3,22 @@ import type { RestHandler } from 'msw';
 type ArrayElementType<T extends ReadonlyArray<unknown>> =
   T extends ReadonlyArray<infer ArrayElementType> ? ArrayElementType : never;
 
-type OptionType = boolean | string | number;
+export type OptionType = boolean | string | number;
+
+export type OptionRenderType = 'text' | 'number' | 'boolean' | 'select';
 
 export type Options<T extends OptionType = OptionType> = Record<
   string,
-  {
-    options: T[];
-    defaultValue: T;
-    selectedValue?: T;
-  }
+  | {
+      options?: T[];
+      selectedValue?: T;
+    } & (
+      | { defaultValue: T; type?: OptionRenderType }
+      | {
+          defaultValue?: T;
+          type: OptionRenderType;
+        }
+    )
 >;
 
 export type ConvertedOptions<T extends Options = Options> = {